feat: Phase 5B - offline-first sync pull from Supabase
- Add PullDtos.kt: @Serializable DTOs for all 6 tables with FK-order pull
- Add PullWorker.kt (HiltWorker): GET /rest/v1/<tabla> for each table in FK
  dependency order; upsertAll into Room; diagnostic Timber logs with row counts
- Add PullTrigger.kt: singleton enqueues PullWorker with ExistingWorkPolicy.REPLACE
- Add upsertAll() to all 6 DAOs (OnConflictStrategy.REPLACE) + contarTodos()
- Add kotlin.serialization plugin to libs.versions.toml and build.gradle.kts
- Change precio from Double to Int across full chain (ProductoEntity, ProductoConStock,
  ProductoRepository, ProductoViewModel, ProductosListScreen); keyboard to Number
- Add Room MIGRATION_3_4: recreate productos table with INTEGER precio,
  FK constraints, no DEFAULT on columns (matches Room schema expectations)
- Trigger PullWorker on app start (StockFlowApp) and after login (AuthRepository)

Fix root causes of pull not showing data:
1. Missing kotlin serialization plugin -> serializers not generated at runtime
2. precio type mismatch: Supabase numeric(12,2) returned 20000.00 as JSON number
3. Room migration had wrong DEFAULT values and missing FK constraints

Validated on physical device: inventory visible per account, multi-tenant isolation confirmed.

+++ app/src/main/java/cl/stockflow/warehouse/data/local/entity/ProductoEntity.kt
@@ -32,7 +32,7 @@ data class ProductoEntity(
     val nombre: String,
     val descripcion: String? = null,
     val sku: String? = null,
-    val precio: Double = 0.0,
+    val precio: Int = 0,
     val stock_minimo: Int = 0,
     // stock NO se almacena — siempre se calcula via MovimientoEntity
     val synced: Boolean = false,
